requirements.txt
six
Django
Django-cors-headers
Django-rest-auth
Djangorestframework
Django-allauth
gunicorn
Django-heroku
psycopg2